revert: for vercel compatibity shifts back to monolithic structure
